feat: use new component in facility list
